requirements.txt
Flask==2.3.3
sentence-transformers==2.6.1
faiss-cpu==1.7.4
numpy>=1.21
scikit-learn>=1.0
requests>=2.25
